+++ src/time/daycounters/actual360.rs
@@ -10,7 +10,7 @@ pub struct Actual360;
 impl DayCounter for Actual360 {
     //
     //
-    fn day_count(&self, date_start: Date, date_end: Date) -> i64 {
+    fn day_count(&self, date_start: Date, date_end: Date) -> usize {
         day_count(date_start, date_end)
     }
