+++ src/time/daycounters/actualactual.rs
@@ -35,7 +35,7 @@ impl DayCounter for ActualActual {
     //
     //
     //
-    fn day_count(&self, date_start: Date, date_end: Date) -> i64 {
+    fn day_count(&self, date_start: Date, date_end: Date) -> usize {
         day_count(date_start, date_end)
     }
